+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Karafka framework changelog
 
 ## 2.0.38 (Unreleased)
-- [Improvement] Introduce `Karafka::Admin#watermark_offsets` to get low and high watermark offsets values.
+- [Improvement] Introduce `Karafka::Admin#read_watermark_offsets` to get low and high watermark offsets values.
 - [Improvement] Track active_job_id in instrumentation (#1372)
 - [Improvement] Improve `#read_topic` reading in case of a compacted partition, there the offset is beyond the low watermark offset. This should optimize reading and should not go beyond the low watermark offset.
 - [Improvement] Allow `#read_topic` to accept instance settings to overwrite any settings needed to customize reading behaviours.